민주	9:50 도서삭제 / 도서검색시 수정삭제 처리

+++ leehyun.book/src/leehyun/book/book/dao/BookDao.java
@@ -17,5 +17,5 @@ public interface BookDao {
 
 	boolean updateBook(Book book);
 
-	boolean delBook(int isbn);
+	boolean delBook(long isbn);
 }
